requirements.txt
# 机器人框架
graia-ariadne==0.11.7
creart==0.3.0
frozenlist==1.4.0
graia-amnesia==0.7.1
graia-ariadne==0.11.7
graia-broadcast==0.23.1
graiax-silkcoder==0.3.6
h11==0.14.0
python-telegram-bot==20.4
discord.py
wechatpy~=2.0.0a26
Werkzeug==3.0.0
Flask-WTF==1.2.0
# LLM上游库
revChatGPT~=6.8.6
openai~=0.27.8
EdgeGPT==0.12.1
OpenAIAuth>=2.0.0
BingImageCreator~=0.4.2
poe-api==0.5.2
tiktoken~=0.4.0

# TTS相关依赖
azure-cognitiveservices-speech
edge-tts

# 其他依赖
toml~=0.10.2
Pillow>=9.3.0
tinydb~=4.8.0
loguru~=0.7.0
asyncio~=3.4.3
pydantic
markdown~=3.4.3
python-markdown-math~=0.8
pygments~=2.15.1
imgkit~=1.2.3
qrcode~=7.4.2
aiohttp~=3.8.5
urllib3~=2.0.4
requests~=2.31.0
uuid~=1.30
aiocqhttp~=1.4.4
tls-client
python-dateutil~=2.8.2
regex~=2023.6.3
httpx~=0.24.1
Quart==0.17.0
pydub~=0.25.1
httpcore~=0.17.3
